refactor: rename blur processor to blur filter and update related files
- Renamed `blur_processor` to `blur_filter` across the codebase for consistency.
- Updated references in the README and BUILD files to reflect the new naming.
- Added new tests for `GaussianBlurFilter` to ensure functionality and correctness.
- Adjusted existing tests to use the updated filter class, maintaining test integrity.
