+++ src/cpp_accelerator/adapters/compute/cuda/filters/blur_filter.cpp
@@ -1,4 +1,4 @@
-#include "src/cpp_accelerator/adapters/compute/cuda/filters/blur_processor.h"
+#include "src/cpp_accelerator/adapters/compute/cuda/filters/blur_filter.h"
 
 #include <cmath>
 #include <string>
