+++ src/cpp_accelerator/adapters/compute/cuda/filters/blur_filter.h
