+++ src/cpp_accelerator/adapters/compute/cuda/filters/blur_filter_test.cpp
@@ -1,4 +1,4 @@
-#include "src/cpp_accelerator/adapters/compute/cuda/filters/blur_processor.h"
+#include "src/cpp_accelerator/adapters/compute/cuda/filters/blur_filter.h"
 
 #include <gtest/gtest.h>
 #include <vector>
