+++ src/cpp_accelerator/adapters/compute/opencl/filters/blur_filter.cpp
