+++ src/cpp_accelerator/adapters/compute/opencl/filters/grayscale_filter_test.cpp
