+++ src/cpp_accelerator/adapters/compute/opencl/kernels/blur.cl
