+++ src/cpp_accelerator/adapters/compute/opencl/kernels/grayscale.cl
