+++ src/cpp_accelerator/adapters/compute/vulkan/filters/blur_filter.h
