+++ src/cpp_accelerator/adapters/compute/vulkan/filters/grayscale_filter_test.cpp
