+++ src/cpp_accelerator/adapters/compute/vulkan/kernels/blur.comp
