+++ src/cpp_accelerator/adapters/compute/vulkan/kernels/grayscale.comp
