--- src/cpp_accelerator/infrastructure/BUILD
@@ -1,6 +1,0 @@
-# Infrastructure layer - adapters and external dependencies
-
-# Subdirectories:
-# - cuda/: CUDA implementations
-# - image/: Image I/O adapters
-# - config/: Configuration management
